==> backend/1_MRT_bus_accessibility.ipynb <==
import requests
import pandas as pd
import geopandas as gpd
import numpy as np
from shapely.geometry import Point
from scipy.spatial import cKDTree
import os



df_resale = pd.read_csv('../data/raw/ResalePrice.csv')
df_resale['address'] = df_resale['block'] + ' ' + df_resale['street_name']

len(df_resale['address'].unique())


### Get coordinates for each address

api_token = "eyJ0eXAiOiJKV1QiLCJhbGciOiJIUzI1NiJ9.eyJzdWIiOiI1M2M4NGU0YmJlMWVlZDhmMDczNDk4ODVmZDExYWRjOSIsImlzcyI6Imh0dHA6Ly9pbnRlcm5hbC1hbGItb20tcHJkZXppdC1pdC1uZXctMTYzMzc5OTU0Mi5hcC1zb3V0aGVhc3QtMS5lbGIuYW1hem9uYXdzLmNvbS9hcGkvdjIvdXNlci9wYXNzd29yZCIsImlhdCI6MTc0Mzc1ODAyNSwiZXhwIjoxNzQ0MDE3MjI1LCJuYmYiOjE3NDM3NTgwMjUsImp0aSI6IkF6YWZjWGxDb2tNb0hmQ1AiLCJ1c2VyX2lkIjozNTA0LCJmb3JldmVyIjpmYWxzZX0.CMD124pML3xaJU45AklBASBYNmojp_wctKoRupiDkQ0"


df_resale

def get_coordinates(address, token):
    url = "https://www.onemap.gov.sg/api/common/elastic/search"
    params = {'searchVal': address, 'returnGeom': 'Y', 'getAddrDetails': 'Y', 'pageNum': 1}
    headers = {'Authorization': token}

    try:
        response = requests.get(url, headers=headers, params=params).json()
        if response['found'] > 0:
            result = response['results'][0]
            return result['LATITUDE'], result['LONGITUDE']
    except requests.exceptions.RequestException:
        pass
    return None


df_resale['latitude'] = None
df_resale['longitude'] = None


unique_addresses = df_resale['address'].unique()
address_coords = {}

for address in unique_addresses:
    coords = get_coordinates(address, api_token)
    if coords:
        address_coords[address] = coords

df_resale['latitude'] = df_resale['address'].map(lambda x: address_coords.get(x, (None, None))[0]).astype(float)
df_resale['longitude'] = df_resale['address'].map(lambda x: address_coords.get(x, (None, None))[1]).astype(float)

### Generate distance to MRT station
Using haversin formula


import numpy as np
from math import radians, cos, sin, sqrt, atan2

bus_stops = gpd.read_file("../data/raw/BusStopLocation_Nov2024/BusStop.shp")
MRT_stops = pd.read_csv("../data/raw/MRT Stations.csv")

def haversine(lat1, lon1, lat2, lon2):
    R = 6371000  # radius of Earth in meters
    phi1, phi2 = radians(lat1), radians(lat2)
    dphi = radians(lat2 - lat1)
    dlambda = radians(lon2 - lon1)
    a = sin(dphi/2)**2 + cos(phi1)*cos(phi2)*sin(dlambda/2)**2
    return 2*R*atan2(sqrt(a), sqrt(1 - a))

from pyproj import Transformer

transformer = Transformer.from_crs("EPSG:3414", "EPSG:4326", always_xy=True)

def svy21_to_wgs84(easting, northing):
    lon, lat = transformer.transform(easting, northing)
    return lat, lon


bus_stops['x_coord'] = bus_stops.geometry.apply(lambda geom: geom.x)
bus_stops['y_coord'] = bus_stops.geometry.apply(lambda geom: geom.y)
bus_stops['Latitude'], bus_stops['Longitude'] = zip(*bus_stops.apply(lambda row: svy21_to_wgs84(row['x_coord'], row['y_coord']), axis=1))

def calculate_nearest_stops(row, mrt_stops, bus_stops):
    hdb_lat, hdb_lon = row['latitude'], row['longitude']
    if pd.isnull(hdb_lat) or pd.isnull(hdb_lon):
        return pd.Series([None, None])

    mrt_distances = mrt_stops.apply(lambda x: haversine(hdb_lat, hdb_lon, x['Latitude'], x['Longitude']), axis=1)
    nearest_mrt_distance = mrt_distances.min()

    bus_distances = bus_stops.apply(lambda x: haversine(hdb_lat, hdb_lon, x['Latitude'], x['Longitude']), axis=1)
    nearest_bus_distance = bus_distances.min()

    return pd.Series([nearest_mrt_distance, nearest_bus_distance])


unique_addresses = df_resale[['latitude', 'longitude']].drop_duplicates().reset_index(drop=True)
unique_addresses[['nearest_mrt_distance', 'nearest_bus_distance']] = unique_addresses.apply(calculate_nearest_stops, axis=1, mrt_stops=MRT_stops, bus_stops=bus_stops)

address_to_nearest_stops = unique_addresses.set_index(['latitude', 'longitude'])[['nearest_mrt_distance', 'nearest_bus_distance']].to_dict('index')

df_resale['nearest_mrt_distance'] = df_resale.apply(lambda row: address_to_nearest_stops.get((row['latitude'], row['longitude']), {}).get('nearest_mrt_distance'), axis=1)
df_resale['nearest_bus_distance'] = df_resale.apply(lambda row: address_to_nearest_stops.get((row['latitude'], row['longitude']), {}).get('nearest_bus_distance'), axis=1)


df_resale[['nearest_mrt_distance', 'nearest_bus_distance']] = df_resale.apply(calculate_nearest_stops, axis=1, mrt_stops=MRT_stops, bus_stops=bus_stops)

df_resale

df_resale.to_csv('../data/cleaned/resale_price_cleaned_0.csv', index=False)

geospatial_df = df_resale[['street_name', 'latitude', 'longitude', 'nearest_mrt_distance', 'nearest_bus_distance']].drop_duplicates()
geospatial_df

#geospatial_df.to_csv('../data/cleaned/hdb_geospatial.csv', index=False)
