==> backend/2_resale_clean.ipynb <==
import pandas as pd

df_resale = pd.read_csv('../data/cleaned/resale_price_cleaned_0.csv')
df_resale

# Convert remaining_lease into numeric variables

def convert_to_years(age_str):
    parts = age_str.split()
    years = int(parts[0])
    months = 0
    if len(parts) > 2:
        months = int(parts[2])
    return round(years + (months / 12),2)

df_resale['remaining_lease'] = df_resale['remaining_lease'].apply(convert_to_years)

# Convert flat types to numerical values

flat_mapping = {
    '1 ROOM': 1,
    '2 ROOM': 2,
    '3 ROOM': 3,
    '4 ROOM': 4,
    '5 ROOM': 5,
    'EXECUTIVE': 6,
    'MULTI-GENERATION': 7,
}

df_resale['flat_type'] = df_resale['flat_type'].map(flat_mapping)
df_resale['storey_range'] = df_resale['storey_range'].str[:2].astype(int)

df_resale = df_resale.drop(columns=['town','block','address'])
df_resale

df_resale.to_csv('../data/cleaned/resale_price_cleaned_1.csv', index=False)
