==> backend/4_macro_features.ipynb <==
import pandas as pd
import numpy as np

df_resale = pd.read_csv('../data/cleaned/resale_price_cleaned_2.csv')
df_resale['month'] = pd.to_datetime(df_resale['month'], format='%Y-%m')
df_resale['month'] = df_resale['month'].dt.strftime('%Y-%m')

### Interest rate

Choose Singapore Overnight Rate Average (SORA) as the indicator for interest rate/

df_interest = pd.read_csv("../data/raw/interest_rate.csv",skiprows=10)
df_interest = df_interest[:8]
df_interest = df_interest.iloc[[6]]

df_interest = df_interest.set_index('Data Series').T.reset_index()
df_interest.columns = ['month', 'interest_rate']
df_interest['month'] = pd.to_datetime(df_interest['month'])
df_interest['month'] = df_interest['month'].dt.strftime('%Y-%m')
df_interest.replace('na', np.nan, inplace=True)
df_interest = df_interest.dropna(subset=['month'])
df_interest

# Join with resale dataframe
df_resale = df_resale.merge(df_interest, on="month", how="left")
df_resale

### Foreign Exchange Rate

df_fxrate = pd.read_csv("../data/raw/fxrate.csv", skiprows=9).head(1)

df_fxrate.replace('na', np.nan, inplace=True)
df_fxrate = df_fxrate.dropna(axis=1)

df_fxrate = df_fxrate.melt(var_name="month", value_name="fx_rate")[1:]
df_fxrate

# Convert month to "YYYY-MM" date format
df_fxrate["month"] = df_fxrate["month"].str.strip()
df_fxrate["month"] = pd.to_datetime(df_fxrate["month"], format='%Y-%m', errors="coerce")
df_fxrate = df_fxrate.dropna(subset=["month"])

# Join with resale dataframe
df_resale = df_resale.merge(df_fxrate, on="month", how="left")
df_resale

### Inflation Rate

The inflation rate is calculated using the CPI.

df_CPI = pd.read_csv("../data/raw/CPI_Base2019_Monthly.csv")
df_CPI.info()
df_CPI = df_CPI.head(1)
df_CPI = df_CPI.T
df_CPI = df_CPI.reset_index() # Convert original column names into the first column

df_CPI = df_CPI.iloc[1:].reset_index(drop=True)
df_CPI.rename(columns={"index":"Time", df_CPI.columns[1]:"CPI (base 2019)"}, inplace = True)
df_CPI.head()


Data cleaning
- In this step, we filter the dataset to include only the rows where the 'Indicator Name' is "Inflation, consumer prices (annual %)". We then reshape the dataframe using the melt function to transform the year columns into rows, with 'Year' as a variable and 'Inflation Rate' as the value. After that, we select only the 'Year' and 'Inflation Rate' columns and drop any rows with missing values.


df_CPI["Time"] = pd.to_datetime(df_CPI["Time"], format='%Y%b').dt.to_period('M')
df_CPI['Time'] = df_CPI['Time'].dt.strftime('%Y-%m')
df_CPI["CPI (base 2019)"] = df_CPI["CPI (base 2019)"].astype(float)

df_CPI.info()

## Analysis
Since inflation is stationary, its statistical properties—such as mean and variance—remain stable over time, indicating that past data can reliably inform future expectations. We can use the inflation level to measure and forecast price index across time, using 2024-12 (dataset's latest) as the base year.

df_CPI

cpi_base = df_CPI.loc[df_CPI['Time'] == '2024-12', 'CPI (base 2019)'].iloc[0]
df_CPI['CPI (base 2024-12)'] = (df_CPI['CPI (base 2019)'] / cpi_base) * 100

# Compute month-over-month inflation rate
df_CPI = df_CPI.sort_values('Time', ascending=True)
df_CPI['inflation_rate (x100)'] = df_CPI['CPI (base 2024-12)'].astype(float).pct_change() * 100

df_CPI = df_CPI[['Time','inflation_rate (x100)']]

df_resale = df_resale.merge(df_CPI, left_on="month", right_on="Time", how="left")
df_resale.drop(columns=['Time'], inplace=True)
df_resale

df_resale.to_csv('../data/cleaned/resale_price_cleaned_3.csv', index=False)
