==> backend/6_private properties.ipynb <==
import pandas as pd
import numpy as np

df_resale = pd.read_csv('../data/cleaned/resale_price_cleaned_4.csv')

df_pri = pd.read_csv('../data/raw/avail_vac_priv_res_prop.csv', skiprows=9).iloc[3:4,:]
df_pri = pd.melt(df_pri, id_vars=['Data Series'], var_name='quarter', value_name='priv_prop')
df_pri.drop(columns=['Data Series'], inplace=True)
df_pri['quarter'] = df_pri['quarter'].str.replace('Q', '').str.replace(' ', '')
df_resale['quarter'] = pd.to_datetime(df_resale['month']).dt.to_period('Q').astype(str).str.replace('Q', '')
df_resale = df_resale.merge(on = 'quarter', right = df_pri, how = 'left')
df_resale.drop(columns=['quarter'], inplace=True)

df_resale.dropna(subset=['resale_price'], inplace=True)

df_resale.drop(columns=['avg_household_income_y'], inplace=True)
df_resale.rename(columns={'avg_household_income_x': 'avg_household_income'}, inplace=True)

df_resale = df_resale.drop(columns=[
    'total_unemployment_rate_y', 
    'resident_unemployment_rate_y'
])

df_resale = df_resale.rename(columns={
    'total_unemployment_rate_x': 'total_unemployment_rate',
    'resident_unemployment_rate_x': 'resident_unemployment_rate'
})

df_resale.to_csv('../data/cleaned/resale_price_cleaned_final.csv', index=False)
