==> backend/7_Normalization.ipynb <==
import pandas as pd
import numpy as np

df = pd.read_csv('../data/cleaned/resale_price_cleaned_final.csv')


df.columns

df_normalized_clean = df.copy()

def normalize(col):
    return (col - col.min()) / (col.max() - col.min()) if col.max() != col.min() else col

df_normalized_clean['month'] = pd.to_datetime(df_normalized_clean['month'])
df_normalized_clean['month_num'] = df_normalized_clean['month'].dt.month.astype(float)
df_normalized_clean = df_normalized_clean.drop(columns=['month'])

df_normalized_clean

df_normalized_clean.info()

columns_to_normalize = [
    'flat_type', 'storey_range', 'floor_area_sqm', 'remaining_lease', 'resale_price',
    'lat', 'lon', 'nearest_mrt_distance', 'nearest_bus_distance', 'education_score',
    'shopping_score', 'food_score', 'recreation_score', 'healthcare_score',
    'inflation_rate (x100)', 'building_age_2025', 'total_unemployment_rate',
    'resident_unemployment_rate', 'interest_rate', 'fx_rate', 'avg_household_income',
    'year', 'priv_prop', 'month_num', 'CPI (base 2024-12)'
]

df_normalized_clean[columns_to_normalize] = df_normalized_clean[columns_to_normalize].apply(normalize)

df_normalized_clean.to_csv('../data/cleaned/resale_price_normalized.csv', index=False)
