==> backend/9_Standardization.ipynb <==
import pandas as pd
import numpy as np

df = pd.read_csv('../data/cleaned/resale_price_cleaned_final.csv')

df_standardized_clean = df.copy()

def standardize(col):
    return (col - col.mean()) / col.std()

df_standardized_clean['month'] = pd.to_datetime(df_standardized_clean['month'])
df_standardized_clean['month_num'] = df_standardized_clean['month'].dt.month.astype(float)
df_standardized_clean = df_standardized_clean.drop(columns=['month'])

columns_to_standardize = [
    'flat_type', 'storey_range', 'floor_area_sqm', 'remaining_lease', 'resale_price',
    'lat', 'lon', 'nearest_mrt_distance', 'nearest_bus_distance', 'education_score',
    'shopping_score', 'food_score', 'recreation_score', 'healthcare_score',
    'inflation_rate (x100)', 'building_age_2025', 'total_unemployment_rate',
    'resident_unemployment_rate', 'interest_rate', 'fx_rate', 'avg_household_income',
    'year', 'priv_prop', 'month_num', 'CPI (base 2024-12)'
]

df_standardized_clean[columns_to_standardize] = columns_to_standardize[columns_to_standardize].apply(standardize)

df_normalized_clean.to_csv('../data/cleaned/resale_price_standardized.csv', index=False)
